backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1

# LangChain + Anthropic
langchain>=0.3.14,<0.4
langchain-anthropic>=0.3.4,<0.4
langchain-core>=0.3.31,<0.4
langchain-community>=0.3.14,<0.4

# Vector DB
chromadb>=0.6.3,<0.7
langchain-chroma>=0.2.2,<0.3

# Embeddings
sentence-transformers>=3.3.1,<4

# SSE streaming
sse-starlette==2.2.1

# Rate limiting
slowapi==0.1.9
